# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EQUHUB API
  description: Полная спецификация API для платформы EQUHUB
  version: 1.0.0
servers:
  - url: http://localhost:80/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /auth/register:
    post:
      summary: Регистрация пользователя
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: Успешная регистрация
        '400':
          description: Ошибка валидации
  /auth/login:
    post:
      summary: Вход пользователя
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
  /auth/logout:
    post:
      summary: Выход пользователя
      tags: [Auth]
      security: [bearerAuth]
      responses:
        '200':
          description: Успешный выход
  /auth/refresh:
    post:
      summary: Обновление токена
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Токен обновлён
...